Update category.js
This code segment defines a virtual field id for the categorySchema schema. The id field is generated by converting the _id field to a hexadecimal string. Additionally, it sets the toJSON option to include virtuals, ensuring that the virtual id field is included when converting the schema instance to JSON.

As well as I entered export part.